Add OTLP/gRPC end-to-end tests using mock collector
## Motivation

In addition to unit tests that orchestrate the components manually, we added some end-to-end tests for the OTLP/HTTP exporter flows that use a mock server and only public APIs (the bootstraps), so we can confirm that they work as intended. We didn't have these for the OTLP/gRPC configuration because the mock collector was buried in a different test target at the time. We ended up missing the fact that the public API wasn't running the exporter (since only the gRPC exporter has a meaningful run method for its connection pool).

We have since restructured the package and now we _do_ have the mock gRPC collector to hand in the test target, so there's nothing stopping us completing the matrix of end-to-end tests to include a configuration for each signal that uses the OTLP/gRPC exporter.

## Modifications

- Add recording logs service to mock OTLP/gRPC collector
- Add OTLP/gRPC end-to-end tests using mock collector

## Result

End-to-end tests now include the use of public API flow with the OTLP/gRPC exporter.