Re-generate files with swift-protobuf 1.31.0 (#339)
## Motivation

Building with Swift OTel with the newly released Swift Protobuf 1.31.0`
produced a couple of instances of this build warning:

> warning: 'init(dictionaryLiteral:)' is deprecated: Please regenerate
your .pb.swift files with the current version of the SwiftProtobuf
protoc plugin. [#DeprecatedDeclaration]

We recently configured our CI to treat warnings as errors so it started
failing unit test jobs.

As suggested, these warnings are fixed by re-generating.

## Modifications

- Re-generate Swift files based on latest Swift Protobuf release

## Result

The project compiles without warnings and therefore makes the unit tests
pass again on CI.