Separate build from tests in CI workflow (#343)
## Motivation

When debugging CI test failures, the build and test output are currently
grouped under a single log section in the GitHub Actions UI, which,
because it contains all the build output, can be slow to load.

## Modifications

Build the package and tests in a separate step, before running the
tests.

## Result

CI logs are easier to view.