Use container jobs instead of Swiftly in CI (#344)
## Motivation

The CI run times are a little unpredictable and it seems that a variable
amount of time is spent installing Swift from Swiftly. In the best case
this is still quite some time and time we needn't spend if we used the
Swift container images.

## Modifications

Move CI to use the `swift` and `swiftlang/nightly-swift` images.

## Result

Faster, less variable CI.