Update benchmarks for new API and add CI for main (#345)
## Motivation

The previous benchmarks no longer compile and we don't currently have
benchmarks for the new APIs or CI on main.

## Modifications

- Update the benchmarks to use the new API with benchmarks for
`withSpan`
- Move benchmarks to the root package behind an environment variable
- Add CI to check for benchmark regressions on pull requests

## Result

- Can benchmark and improve code locally
- CI will check for regressions

## Notes

The moving of the benchmarks to the root package is inspired by
valkey-swift, and it has many benefits:

- No more hopping between the packages (since Xcode cannot have both the
root and the subdirectory package open concurrently).
- Ability to benchmark code that has package access.

I have some performance improvements waiting, but it would be good to
get this in to establish a baseline.

---------

Co-authored-by: Moritz Lang <16192401+slashmo@users.noreply.github.com>