Remove OTLPExporterConfiguration.insecure config option (#348)
## Motivation

There's a bunch of confusion that often arises from the use of
`OTLPExporterConfiguration.insecure: Bool` since it has non-obvious
semantics in the OTel spec. It is only used for OLTP/gRPC exporters, and
silently ignored for OTLP/HTTP. It also is only used when the endpoint
does not contain a scheme. The spec states that SDKs can choose to not
support this option if it doesn't fit well with the underlying gRPC
client library[^1] and, given our library requires us to be explicit
about this, I think we should take the opportunity to remove the
ambiguity from the API surface.

## Modifications

Remove `OTLPExporterConfiguration.insecure` config option

## Result

Less confusing API.

## Note

This is an API breaking change, which we should call out in the release
notes, and should be done _before_ 1.0.

[^1]:
https://opentelemetry.io/docs/specs/otel/protocol/exporter/#specify-protocol