Combine coverage from matrix jobs of all Swift versions (#351)
## Motivation

We have a coverage report generated at the end of the unit test
pipeline, but only for Swift 6.1. However, we have a bunch of tests that
make use of features that are only in Swift Testing from 6.2 toolchains,
so these currently do not contribute to the coverage result. Now Swift
6.2 is released, it seems reasonable that we should include these too.

## Modifications

- Combine coverage from matrix jobs of all Swift versions

## Result

Coverage result should now include tests that require Swift 6.2.