Benchmark CI should only add PR comment for non-fork PRs (#355)
## Motivation

The benchmark PR can only comment on PRs made from non-fork branches
because the workflow is `on: pull_request`. I've researched ways around
it but getting it to work with fork PRs would be complicated to do in a
secure way (see
https://securitylab.github.com/resources/github-actions-preventing-pwn-requests/).

## Modifications

For now, we should juts only try and make the comment on the PR if the
PR is from a non-fork.

## Result

Benchmark CI won't be falsely red.