Add support for histogram min and max values (#356)
## Motivation

The internal histogram type only keeps track of the running total and
bucket counts. It does not keep track of the min and max value seen.
When exporting over OTLP, it therefore only includes the `sum` and
bucket counts. The `min` and `max` fields are left unset (which results
in 0 values).

## Modifications

- Add support for histogram min and max values
- Update tests to check for min and max histogram values

## Result

OTLP export of histogram metrics includes min and max values.