Add DocC curation for otelTraceID ServiceContext property (#361)
## Motivation

Symbols that haven't been _curated_ in the DocC bundle hang out in an
automatic section. In the case of the trace ID context property, it gets
lumped into _extensions on other modules_, but it'd be nice to have it
categorised in the sidebar, since it's the only thing that is not.

## Modification

Curate `ServiceContext.otelTraceID` in DocC

## Result

Docs better organised.