Add workaround for trait-based dependencies in broken 6.2.x toolchains (#357)
## Motivation

We had a workaround in our package manifest for the use of
`.when(traits:)` not using its argument as a boolean-OR. The good news
is that this bug has been addressed and the fix available in the 6.2.0
release. The bad news is that there was still an issue with resolving
dependencies that themselves had trait-conditional transitive
dependencies, and the fix for that did _not_ make the Xcode 26.0
release.

The outcome is that if you were to try to compile a project that depends
on Swift OTel with _only_ the `OTLPHTTP` trait, it will fail to resolve.
We have such an example in our `Examples/` directory:

```console
% rm -rf .build Package.resolved && swift build
...
error: 'swift-otel': product 'SwiftProtobuf' required by package 'swift-otel' target 'OTel' not found in package 'swift-protobuf'.
```

The fix is included in the nightly main, and I have verified that works.
It also has been cherry-picked into 6.2.1 here:
https://github.com/swiftlang/swift-package-manager/pull/9141. So it is
just 6.2.0 that's the problem.

Fortunately, all our trait-conditional dependencies are only used with
`internal import`, which means we can add a workaround without having to
make a future API-breaking change.

We can just unconditionally include these dependencies on the known
problematic toolchain (6.2.0).

## Modifications

- Add 6.2.0-specific workaround for trait-conditional dependencies

## Result

Packages that depend on Swift OTel with only `OTLPHTTP` trait will now
compile with Xcode 26.0 toolchains.