Improve runtime performance of withSpan (#360)
## Motivation

Now we have benchmarks in place I took a look at the cost of `withSpan`
both for the `alwaysOn` and the `alwaysOff` samplers. There was some
room for improvement.

Because of the `nonmutating set` requirement of `protocol Span`, each of
the mutable properties is pushed into a locked value box. However, this
means we perform a heap allocation per property. This isn't necessary,
and bundling up all the mutable properties into a private struct, in a
single locked value box, reduces the allocations.

Furthermore, when the sampler is dropping the span, we are paying more
allocations than necessary because we are doing a bunch of up-front work
that is later discarded.

## Modifications

- Reduce allocations by bundling state behing one locked box
- Refactor to avoid allocations for dropped spans
- Add fast path for dropped spans

## Result

- Performance improved for withSpan generally.
- Performance of withSpan significantly improved for dropped spans.