Add API breakage pipeline using custom script (#365)
## Motivation

Unfortunately using `swift package diagnose-api-breaking-changes` isn't
working for this package right now. It looks to be an issue with the way
it's passing the include flags for one of the calls to
`swift-api-digester`. For the time being though, we'd really benefit
from a pipeline that does check for breakages.

## Modification

- Add `scripts/check-api-breakage.sh` which drives `swift-api-digester`
manaually
- Add CI job to run the check

## Result

PRs should show API breakages