Remove old environment type and resource detection (#366)
## Motivation

We have an `OTelEnvironment` type, which these days is internal and not
required through paths from the public API. It used to be used for
configuration and resource detection. The configuration API is now
implemented without this type and resource detection is not yet
available in the public API and, when we make it available, we likely
want to do it differently.

It was still being constructed on the API-reachable path though, and
then never used, and this construction causes an error when there are
environment variables that differ only in their case, as raised in #364.

## Modifications

- Remove `OTelEnvironment` and all its uses

## Result

- Simpler code base, removing code that we don't use now.
- Fixes #364.