Log bootstrap at info and downgrade all other logging to debug (#368)
## Motivation

Our startup logging is noisy and inconsistent. For logging it's very
useful and only verbose when there's a useful information to convey to
the user. But we also log a bunch of lifecycle events at info. On
reflection, I think this is a sub-par experience out of the box.

## Modifications

- Add an info level bootstrap message for traces and metrics, like we
have for logging.
- Downgrade all other logging to debug.

## Result

### Before

```
2025-09-29T09:34:57+0100 info swift-otel: component=bootstrap [OTel] Bootstrapping logging system with OTLP/HTTP+Protobuf exporter.
---
Only Swift OTel diagnostic logging will use the console logger.

If you require console logging for local development, use the
console logs exporter, which can be enabled using the
following configuration:

    config.logs.exporter = .console

Or, run your process with the following environment variable:

    OTEL_LOGS_EXPORTER=console

If you require logs to go to both the console and another
exporter, manually bootstrap the logging subsystem with a
multiplex log handler. See the documentation of
`makeLoggingBackend` for details.
---
2025-09-29T09:34:57+0100 info swift-otel: component=OTelBatchLogRecordProcessor [OTel] Starting.
2025-09-29T09:34:57+0100 info swift-otel: component=OTelTracer [OTel] Starting.
2025-09-29T09:34:57+0100 info swift-otel: component=OTelPeriodicExportingMetricsReader interval=3.0 seconds [OTel] Started periodic loop.
```

### After

```
2025-09-29T09:34:03+0100 info swift-otel: component=bootstrap [OTel] Bootstrapping logging system with OTLP/HTTP+Protobuf exporter.
---
Only Swift OTel diagnostic logging will use the console logger.

If you require console logging for local development, use the
console logs exporter, which can be enabled using the
following configuration:

    config.logs.exporter = .console

Or, run your process with the following environment variable:

    OTEL_LOGS_EXPORTER=console

If you require logs to go to both the console and another
exporter, manually bootstrap the logging subsystem with a
multiplex log handler. See the documentation of
`makeLoggingBackend` for details.
---
2025-09-29T09:34:03+0100 info swift-otel: component=bootstrap [OTel] Bootstrapping metrics system with OTLP/HTTP+Protobuf exporter.
2025-09-29T09:34:03+0100 info swift-otel: component=bootstrap [OTel] Bootstrapping instrumentation system with OTLP/HTTP+Protobuf exporter.
```