Prepare docs and examples for 1.0.0 release (#358)
## Motivation

We're close to tagging the 1.0 release. It would be good if the repo, at
this tag, had all the examples and documentation updated as if the tag
was available. For that, we'll need to update them just prior to
tagging.

## Modifications

- Update docs to use `from: "1.0.0"`
- Update examples to use `from: "1.0.0"`

## Result

Docs and examples show how to depend on Swift OTel in a post-1.0 world.