Bump protobuf dependency to 1.31.0 (#375)
## Summary

Fixes https://github.com/swift-otel/swift-otel/issues/374.

When a new version of the Swift protobuf generator uses new features not
present in an older version of the runtime library, the runtime
dependency needs to be bumped, otherwise adopters who resolve the
minimum version end up with a build error.

## Test Plan

N/A